+++ src/main/java/com/haibao/business/domain/enums/GloryTypeCode.java
@@ -5,20 +5,20 @@ package com.haibao.business.domain.enums;
  * Created by haibao on 2018/1/22.
  */
 public enum GloryTypeCode {
-    PERSONAL((byte) 0, "个人奖"),
-    TEAM((byte) 1, "团体奖");
+    PERSONAL( 0, "个人奖"),
+    TEAM( 1, "团体奖");
 
 
 
-    GloryTypeCode(byte code, String desc) {
+    GloryTypeCode(int code, String desc) {
         this.code = code;
         this.desc = desc;
     }
 
-    private byte code;
+    private int code;
     private String desc;
 
-    public byte code() {
+    public int code() {
         return code;
     }
