+++ src/main/java/com/haibao/business/domain/enums/TermCode.java
@@ -5,19 +5,19 @@ package com.haibao.business.domain.enums;
  * Created by haibao on 2018/1/22.
  */
 public enum TermCode {
-    FIRST((byte) 0, "第一学期"),
-    SECOND((byte) 1, "第二学期");
+    FIRST( 0, "第一学期"),
+    SECOND( 1, "第二学期");
 
 
-    TermCode(byte code, String desc) {
+    TermCode(int code, String desc) {
         this.code = code;
         this.desc = desc;
     }
 
-    private byte code;
+    private int code;
     private String desc;
 
-    public byte code() {
+    public int code() {
         return code;
     }
